Task 7 Added
Implemented a basic leveling system that adds a finite amount of rows and an infinite amount of extra speed to the ball each time the player removes all the bricks. The level number is displayed in the center of the screen. I  also increased the speed of the paddle because I thought it felt much slower than most other Breakout games.